refactor: refactor semantic tests
