+++ src/test/scala/org/bitbucket/inkytonik/cooma/test/SemanticTests.scala
@@ -1,0 +1,35 @@
+package org.bitbucket.inkytonik.cooma.test
+
+import org.bitbucket.inkytonik.cooma.CoomaParserSyntax.{ASTNode, Program}
+import org.bitbucket.inkytonik.cooma.{ReferenceDriver, SemanticAnalyser}
+import org.bitbucket.inkytonik.kiama.relation.Tree
+import org.bitbucket.inkytonik.kiama.util.{StringSource, Tests}
+
+trait SemanticTests extends Tests {
+
+    val driver = new ReferenceDriver
+    val config = {
+        val newConfig = driver.createConfig(Seq())
+        newConfig.verify()
+        newConfig
+    }
+
+    def test(name : String, expression : String, expectedMessages : String) : Unit =
+        test(name) {
+            runAnalysis(expression.stripMargin) shouldBe expectedMessages.stripMargin
+        }
+
+    def runAnalysis(expression : String) : String = {
+        val messages =
+            driver.makeast(StringSource(expression), config) match {
+                case Left(ast) =>
+                    val tree = new Tree[ASTNode, Program](ast)
+                    val analyser = new SemanticAnalyser(tree)
+                    analyser.errors
+                case Right(messages) =>
+                    messages
+            }
+        driver.messaging.formatMessages(messages)
+    }
+
+}
